tdf#107020: Reintroduce uses of I32S, I64S
...that had been removed with 1b254cd2a1c5950bef59d97e7557f364835ddb37 "There
should be no need to distinguish between I32S and I64S" (which this commit
effectively reverts).  I had erroneously assumed that the arguments to I32S and
I64S would only be of (signed) integer types, and that it would be irrelevant
that they are truncated to sal_Int32 resp. sal_Int64.  However, at least
oox/source/export/drawingml.cxx called I32S with double values (for which there
is an OUString::number overload too), so the resulting strings could be rather
different.

Change-Id: Ib3887269fad0764740bc9ed3be91f67696a23b55
Reviewed-on: https://gerrit.libreoffice.org/37040
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>