tdf#107474 sw: fix crash when deleting header with active SdrObject
The problem is that the SdrView's text edit mode remains active when
the header is deleted, which deletes the active SdrObject that is
anchored in the header.

Then the next call to SdrBeginTextEdit() will call SdrEndTextEdit()
but the mxTextEditObj WeakReference is null then and so the cleanup
code that resets the SdrOutliner's UndoManager is not called,
but then the SdrOutliner is deleted.

What's really horrible about this is that while text editing is active,
both SdrObjEditView and ImpEditEngine consider themselves the
exclusive owner of SwDoc's UndoManager.

(regression from 203faed0ebcfaab1d171ec4f6e5f20215998acab
 and 5cf8ff8a8720a45b6b51b3bb5831b3411addcd2b)

Change-Id: I17c5165a03d2de1eaba8a335dfbfded56573f8a2