Revert "tdf#107204: Make sure we always do Unicode Bidi"
This reverts commit 2672f8ce13b9351401e8905e69c45ec379058a06, as it
breaks the testTableKeep part of CppunitTest_sw_ww8export on the
Jenkins_Callgrind tinderbox (and also for me locally) and there is no
fix in progress to address that.

See the bug for a comparision of the old/new layout for the problematic
testcase.

Khaled says it's OK to push this revert till it's clear how to fix the
bug without side-effects.

Change-Id: I669de98ffa8b519024e332ae468808d7e7243df2