tdf#88555 sw: reduce usage of dynamic_cast in SwFormatsModifyBase::Contains
This is a bad idea as the function is sometimes used to check if a
SwFormat has been deleted, which can happen in Undo.

Replace with ContainsFormat() and IsAlive(), and don't require a
non-const SwFormat parameter so that the dynamic_cast using one
isn't called all the time but only called once where it's safe.

Change-Id: Ib74b79629f5c8ed432a912ada5974a6d816e7d31