tdf#107566 adjust constraints for Add-In function YIELDMAT.
Constraints are adjusted to make Calc behave the same as Excel.
Constraint for rate is less severe than ODFF1.2, but a rate of 0 is
mathematically no problem and Excel allows a rate of 0.
Constraint that Settlement date cannot be before Issue date is neither
mentioned in ODFF1.2 nor in Excel's help text for YIELDMAT, but is applied
by Excel. In reality, a settlement before issue is not possible.

Change-Id: I9c19d33abe8518311606fcee1c9dea32438d61a4
Reviewed-on: https://gerrit.libreoffice.org/37206
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Eike Rathke <erack@redhat.com>