Revert "tdf#96444 Replace chr(20) with '%20' in style uno commands"
This reverts commit fdca83206f002008a15d2fdb036f4dadcee75f93.

This causes styles to not appear as having keyboard
shortcuts in Tools > Customize... > Keyboard, and style
buttons that were added using tdf#106681 to not have
their keyboard shortcuts in their tooltips.

Given that the original bug is a local problem in
MenuBarManager, I prefer to fix it there, rather than
put workarounds all over the place, including those ugly
"%20".

Change-Id: I797f463df2b37ab69f73f357ab9d120697e0ff33