ensure cow_wrapper contents are queried through a const pointer
otherwise the query will call make_unique to make this a unique
copy of the underlying shared font which is a waste if the
contents are going to be the same and the if branch contents
are skipped.

I was tempted to call the Font accessors like GetLanguage, etc.
but a bunch of the accessors have two versions, a const and non-const
one which behave differently so that looks like another rats nest

This relates to the changes of

commit 7ee0714581db97a3cda18c01abf7c93bfe376a44
Date:   Fri Apr 22 23:43:52 2016 +0200

    tdf#62525 vcl: use cow_wrapper for font

Change-Id: Ic0661fccb414b3636308975e265fe5751476e1b8
Reviewed-on: https://gerrit.libreoffice.org/37502
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>