tdf#107773 DOC import: consider page breaks for multi-page floattables
See commit 254ac1f4405b576bd44f2b6a602ce7088e111081 (Related: fdo#61594
SwWW8ImplReader::StartApo: don't always start a frame, 2013-05-13) for
the context, this is one more case where having a multi-page floattable
as a multi-page one is a better option.

The additional condition is that in case there is a break right after
the table, then always import the table as a non-floating one, as there
won't be any wrapping anyway (due to the break).

Change-Id: I3dd4174226cfc1e825c00607221feeed51b7133f
Reviewed-on: https://gerrit.libreoffice.org/37513
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>