Handle reference list as array in IterateParameters(), tdf#58874
This in array context now returns the expexted array for an expression like
=SUBTOTAL(9,OFFSET(A1,ROW(1:3),0,2))
where OFFSET returns an array (svRefList) of references, here
A2:A3
A3:A4
A4:A5
and SUBTOTAL(9,...) sums each of the references independently to return an
array of three sums.

Currently works with SUM, COUNT, COUNT2, AVERAGE, PRODUCT and SUMSQ.
MAX and MIN are not handle in IterateParameters() and need an extra
implementation, or rather move them into IterateParameters() as well?

Change-Id: I03b7d1cccf78c1a831348106432126aec8d1f519