tdf#107998 - svg export: shapes are uncorrect both in Draw and Impress
Start Draw, draw a rectangle, export it to a SVG document.

Open the SVG document with a browser or Inkscape: instead of a
rectangle, you will see a self-crossing polygon.

This issue is due to a clean up commit
(fccf2034d91bdec6950f45404f3e2ae32adbda4b) about string concatenation
which has not taken into account that operations are performed from
right to left.

Change-Id: Ib1c0d872b6bd95abe98bb5e22ef69f6b2d5c163b
Reviewed-on: https://gerrit.libreoffice.org/37899
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Tested-by: Jan Holesovsky <kendy@collabora.com>