Makefile: try to fix downloading externals for build platform
When cross-compiling, it's possible that an external is only used on
the build platform and not the host platform.  Attempt to fix this by
invoking "make fetch" twice; it looks like this should not be
combined with the exising "build-tools" invocation because that needs to
be delayed until everything is fetched.

Change-Id: I42626a30359d54dce3c0a5a8ea07045c099cea59
Reviewed-on: https://gerrit.libreoffice.org/37912
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Michael Stahl <mstahl@redhat.com>