tdf#108056 sw SubtractFlys: add missing expansion of the clip path
With this finally the bugdoc has no more unexpected white lines around
the fly frame. In the non-SubtractFlys case DrawFillAttributes() already
did an expansion of the clip path, but this was overwritten in case the
layout flag was set.

Fix the problem by taking care of this in lcl_SubtractFlys() itself.

Change-Id: Iac91279f8bc19588e763425aff5cb800e793da83
Reviewed-on: https://gerrit.libreoffice.org/38079
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>