tdf#107682 - Repeated images replace correct ones in exported PDF
The problem was due to the buffer acquire methods: in Bitmap::Checksum
(old implementation) Bitmap::AcquireReadAccess is used to get the bitmap
buffer: indeed this method relies on SalBitmap::AcquireBuffer (which is
used in the new implementation) but in case the buffer acquisition
fails, instead of giving up, it tries to update the imp bitmap instance
embedded in the bitmap (see BitmapInfoAccess::ImplCreate).

The solution is to perform this further attemp in Bitmap::Checksum when
the value returned by ImpBitmap::GetChecksum is 0.

Change-Id: Ib901ac941db57756e9a951bacbc573ca206316e0
Reviewed-on: https://gerrit.libreoffice.org/38165
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Marco Cecchetti <mrcekets@gmail.com>