tdf#96256 sw: fix scrolling to cursor after paste
The mnStartAction hack in commit e1ddd5dc3161af07251ba3e8ee26d8fcb5ee5e4c
has a surprising effect on SwViewShell::SizeChgNotify(), as it causes
PageNumNotify() to be called instead of SizeNotify() if pages are
inserted while layouting in GetCharRect().

Now SwCursorShell::UpdateCursor(SCROLLWIN) would no longer scroll to
the cursor position, which is a problem for clipboard paste.

Try to fix this by calling UISizeNotify() manually, which seems to work.

Change-Id: Id08bfe2964564c02fe9d1ef0176939ab3eeb187a